+++ src/ECS/TransformComponent.hpp
@@ -7,29 +7,29 @@ struct TransformComponent : public Component
     Vector2D position;
     Vector2D velocity;
 
-    int width;
-    int height;
+    float width;
+    float height;
     int scale;
     int speed;
 
     TransformComponent()
     {
         position.Zero();
-        width = height = 32;
+        width = height = 32.0f;
         scale = 1;
         speed = 3;
     }
 
     TransformComponent(float x, float y)
     {
-        width = height = 32;
+        width = height = 32.0f;
         scale = 1;
         speed = 3;
         position.x = x;
         position.y = y;
     }
 
-    TransformComponent(float x, float y, int w, int h, int s)
+    TransformComponent(float x, float y, float w, float h, int s)
     {
         speed = 3;
         position.x = x;
@@ -41,14 +41,14 @@ struct TransformComponent : public Component
 
     TransformComponent(float x, float y, int speed)
     {
-        width = height = 32;
+        width = height = 32.0f;
         scale = 1;
         position.x = x;
         position.y = y;
         this->speed = speed;
     }
 
-    TransformComponent(float x, float y, int w, int h, int s, int speed)
+    TransformComponent(float x, float y, float w, float h, int s, int speed)
     {
         position.x = x;
         position.y = y;
